+++ web_v2/requests/get_balance_list.go
@@ -13,6 +13,9 @@ const (
 	// FilterTypeBalanceListAsset - defines if we need to filter the list by asset
 	FilterTypeBalanceListAsset = "asset"
 
+	// FilterTypeBalanceListOwner - defines if we need to filter the list by owner
+	FilterTypeBalanceListOwner = "owner"
+
 	// FilterTypeBalanceListAssetOwner - defines if we need to filter the list by asset owner
 	FilterTypeBalanceListAssetOwner = "asset_owner"
 )
@@ -24,6 +27,7 @@ var includeTypeBalanceListAll = map[string]struct{}{
 var filterTypeBalanceListAll = map[string]struct{}{
 	FilterTypeBalanceListAsset:      {},
 	FilterTypeBalanceListAssetOwner: {},
+	FilterTypeBalanceListOwner:      {},
 }
 
 // GetBalanceList - represents params to be specified by user for getBalanceList handler
@@ -32,6 +36,7 @@ type GetBalanceList struct {
 	Filters struct {
 		Asset      string `fig:"asset"`
 		AssetOwner string `json:"asset_owner"`
+		Owner      string `json:"owner"`
 	}
 	PageParams *db2.OffsetPageParams
 }